OPTIMIZATION: Refactored data structures of the contact storage for improved search speed
Contacts are now stored in memory with HashMaps with the contact uuid as the  HashMap's Key. Contact now has an `id` field for holding contact's unique Uuid instead of identifying contact by index position in a Vec

The in-memory index now identify contacts via Uuid enabling the ability to built once and update index as contact list changes without invalidating the Index.